Refactor attributes input validation (#6535)
* Refactor validate_attributes_input util function

* Move AttributeAssignmentMixin and validate_attributes_input to attribute module
